refactor: replace scrollbar config to rely on defaults
The current neovim config has been using nvim-scrollbar's default
configuration settings since the plugin has been added. The
nvim-scrollbar-config file is not necessary and we can trust
the plugin to have mostly consistent backwards compatibility with
it's default configuration. The change could have been possible before
the introduction of lazy vim but it is a good opportunity now.